+++ index.html
@@ -11,6 +11,6 @@
     />
   </head>
   <body>
-    <script type="module" src="script.js"></script>
+    <script type="module" src="app.js"></script>
   </body>
 </html>
