+++ javascript/creating-keys.js
